+++ src/Presentation/Factory/EventFactory.php
@@ -30,6 +30,7 @@ use Eluceo\iCal\Presentation\Component\Property\Value\DateTimeValue;
 use Eluceo\iCal\Presentation\Component\Property\Value\DateValue;
 use Eluceo\iCal\Presentation\Component\Property\Value\GeoValue;
 use Eluceo\iCal\Presentation\Component\Property\Value\IntegerValue;
+use Eluceo\iCal\Presentation\Component\Property\Value\ListValue;
 use Eluceo\iCal\Presentation\Component\Property\Value\TextValue;
 use Eluceo\iCal\Presentation\Component\Property\Value\UriValue;
 use Generator;
@@ -111,6 +112,10 @@ class EventFactory
             foreach ($event->getAttendees() as $attendee) {
                 yield $this->attendeeFactory->createProperty($attendee);
             }
+        }
+
+        if ($event->hasCategories()) {
+            yield $this->getCategoryProperties($event);
         }
 
         foreach ($event->getAttachments() as $attachment) {
@@ -219,5 +224,15 @@ class EventFactory
         }
 
         return new Property('ORGANIZER', new UriValue($organizer->getEmailAddress()->toUri()), $parameters);
+    }
+
+    private function getCategoryProperties(Event $event): Property
+    {
+        $categories = [];
+        foreach ($event->getCategories() as $category) {
+            $categories[] = new TextValue((string) $category);
+        }
+
+        return new Property('CATEGORIES', new ListValue($categories));
     }
 }